+++ CHANGELOG.md
@@ -45,6 +45,7 @@ v5.4.0
 - Update Session File for Parallel-in-Time (!1516)
 - New version of CreateCollection using basisKey (!1603)
 - Remove unused Domain.cpp and Domain.h file (!1609)
+- Remove useless ReadExpressions and SubstituteExpressions function in sessionReader (!1608)
 
 **CompressibleFlowSolver**
 - Fix AUSM3 Riemann solver and add tests to the CI (!1537)